Homework 1 complete: README with badges, Terraform variables, learning log

+++ Module_1/Homework_1/ingest_to_db.py
@@ -29,7 +29,7 @@ dtype_trip = {'VendorID': 'Int64',
  'cbd_congestion_fee': 'float64'}
 
 dtype_lookup = {
-    "LocationID" : "int32",
+    "LocationID" : "Int64",
     "Borough" : "str",
     "Zone" : "str",
     "service_zone" : "str"
